Added authentication and databases
The 'default' Boot profile will use an in-memory database with Spring Security's default schema.

The 'local' profile will use a PostgreSQL database running on localhost and will populate the database with a clean Timneh schema and seed data each time it's run.
Added docker/postgresql for an easy way to spin up a database with Docker.